third_party/WebKit/LayoutTests/plugins/netscape-plugin-setwindow-size-2-expected.txt

CONSOLE MESSAGE: PLUGIN: NPP_SetWindow: 200 200
This tests that a hidden plugin gets a correct NPP_SetWindow the first time.
